feat: compat old create tools
