+++ packages/generator/generators/repo-next-generator/modern.config.js
@@ -1,0 +1,5 @@
+const { generatorBuildConfig } = require('@scripts/build');
+
+module.exports = {
+  buildConfig: generatorBuildConfig,
+};
