+++ packages/generator/generators/repo-next-generator/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@modern-js/repo-generator",
+  "name": "@modern-js/repo-next-generator",
   "description": "A Progressive React Framework for modern web development.",
   "homepage": "https://modernjs.dev",
   "bugs": "https://github.com/web-infra-dev/modern.js/issues",
@@ -29,10 +29,21 @@
     "test": "jest --passWithNoTests"
   },
   "devDependencies": {
+    "@modern-js/codesmith-utils": "2.6.0",
+    "@modern-js/base-generator": "workspace:*",
     "@modern-js/codesmith": "2.6.0",
-    "@modern-js/repo-next-generator": "workspace:*",
+    "@modern-js/codesmith-api-app": "2.6.0",
+    "@modern-js/generator-common": "workspace:*",
+    "@modern-js/generator-plugin": "workspace:*",
+    "@modern-js/generator-utils": "workspace:*",
+    "@modern-js/module-generator": "workspace:*",
+    "@modern-js/mwa-generator": "workspace:*",
+    "@modern-js/plugin-module-banner": "workspace:*",
     "@scripts/build": "workspace:*",
+    "@scripts/jest-config": "workspace:*",
+    "@types/jest": "^29",
     "@types/node": "^14",
+    "jest": "^29",
     "typescript": "^5"
   },
   "sideEffects": false,
